Fixed delimiter handling in cookie expiration
RFC6265 allows much broader delimiter options than just space in
Set-Cookie expires values.  When these delimiters are -, strtol()
treats numbers as negative.  When they are any of the other thirtyish
values, strtol() won't parse the values at all.

Rather than try to handle all the legal values explicitly, we just skip
over the character following the last parsed token, blindly assuming
that it is a legal delimiter.  If we get to the end and the date is
sane, we assume we did the right thing in the parser above.