Suppressed cookie time zone offsetting for GMT/UTC
If the Set-Cookie header's expires field contains " UTC" or " GMT",
we suppress the timezone offsetting.  Without this, a server in the
Western hemisphere that returns a UTC-referenced cookie that expires
in fewer hours than your local offset from UTC wouldn't get that
cookie back because Siege thought it had expired already.

That is, a server in GMT-7 that returns a GMT-referenced cookie that
expires in an hour wouldn't get that cookie back on the next request
because Siege was subtracting 7 hours from the cookie, and so was
thinking it expired 6 hours ago.

The eastern hemisphere case is less critical, since it would require
shennanigans on the server's side to exploit.